backend/requirements.txt
fastapi
uvicorn
openai
pydantic
pydantic-settings
sqlalchemy
python-dotenv
aiosqlite
